Organize src/ directory.
The files in the src/ directory will be organized based on which
compilers or boards can use them.

The directory src/all/ will be for files that work for all compilers and
boards.

The directory src/arduino/ will be for files that depend on Arduino
(Arduino.h).

Other directories will be created as the need arises.

The only files in the src/ directory will be README.md and
MorseCodeMachine.h.