Organize src/ directory.
As much as possible, a .h file should only expose the functionality implemented
by its .cpp file.

Then MorseCodeMachine.h will combine all the .h files to expose all of the
functionality in the library.

Of course all .h files in the src/arduino/ directory will most of the time
expose the functionality in Arduino.h, but there is not much that can be done
about that.