Add context parameters to the sendMorse functions.
This gives the caller of the sendMorse function the ability to add state to
their function pointers (delayFunction, dotFunction, and dashFunction) without
the use of global variables.

Even though the use of pointers is somewhat discouraged by the Arduino style
guides, I think this is still okay to add to the library because this addition
does not force anyone to use pointers.

Why not just use C++ classes?
If you look at what is available in Arduino, functions seem like a better
fit. Also, there are other ways of managing state and functions without
the complexity of classes and without the complexity of pointers.

void sendMorse(const char message[],
               void (*delayFunction)(void *context),
               void (*dotFunction)(void *context),
               void (*dashFunction)(void *context),
               void *context);

void sendMorse(const String &message,
               void (*delayFunction)(void *context),
               void (*dotFunction)(void *context),
               void (*dashFunction)(void *context),
               void *context);

void sendMorse(const String *message,
               void (*delayFunction)(void *context),
               void (*dotFunction)(void *context),
               void (*dashFunction)(void *context),
               void *context);