Add extras/test.cpp.
Add a small test program in the extras/ directory. The extras/ directory is
guaranteed to be ignored by the Arduino IDE. The test program must be small
so that it will not take up too much space in the library.

The test program will return the number of failures as its error code.