OPTIMIZATION: Optimize char to morse code conversion.
What registers are used by the C compiler?
https://www.microchip.com/webdoc/AVRLibcReferenceManual/FAQ_1faq_reg_usage.html

According to the AVR Libc Reference, arguments to function calls and return
values from functions are stored in registers if possible, instead of putting
them on the stack.

So the charToMorseElements function was changed from
   static unsigned char charToMorseElements(char, int &)
to
   static unsigned int charToMorseElements(char)

Morse code will not be changing much anymore, so I think this kind of thing is
okay to do.