Modify sendMorse.
Modify sendMorse so that messages can have nested angle brackets.

Doing this will be helpful to the convertMorseFromGreek function and future
convertMorseFrom<Language> functions.

For example, one Greek character has the pattern "- - - -", which can be
represented by bracketed english characters such as "<TTTT>", "<MM>", or
"<MTT>". The problem is that when that Greek character is already inside
angle brackets, then generating "<MM>" for that Greek character will create
a message with "<MM>" nested inside another set of angle brackets.

It seems simpler to handle the nesting of angle brackets.