Add accented Latin characters.
The following characters are now supported:
ÀàÄäÅåĄąÆæĆćĈĉÇçĐđÐðÉéÈèĘęĜĝĤĥĴĵŁłŃńÑñÓóÖöØøŚśŜŝŠšÞþÜüŬŭŹźŻż

Pedantic Note
=============
Most of the documentation in the header files, examples, and function
parameters refer to converting non-Latin characters to English.

Now that this library supports more characters based on Latin and not just
English characters based on Latin, you *may* be tempted to update the
documentation to be more precise.

Nope. Not happening.

To be more precise it did happen, and if you want to see it, look at the
"pedantic" branch. It did not really improve anything.